Move `CRSTransform` to `geoarrow_schema` and make title case
