+++ rust/geoarrow-flatgeobuf/src/writer.rs
@@ -3,8 +3,8 @@ use std::io::Write;
 use arrow_schema::Schema;
 use flatgeobuf::{FgbCrs, FgbWriter, FgbWriterOptions};
 use geoarrow_array::GeoArrowType;
-use geoarrow_array::crs::{CRSTransform, DefaultCRSTransform};
 use geoarrow_array::geozero::export::GeozeroRecordBatchReader;
+use geoarrow_schema::crs::{CRSTransform, DefaultCRSTransform};
 use geoarrow_schema::error::{GeoArrowError, GeoArrowResult};
 use geoarrow_schema::{Dimension, Metadata};
 use geozero::GeozeroDatasource;
